Fix a 32-bit-ism.

+++ main.c
@@ -265,11 +265,11 @@ NetInfo net_info[JNB_MAX_PLAYERS];
 
 typedef struct
 {
-	unsigned long cmd;
-	long arg;
-	long arg2;
-	long arg3;
-	long arg4;
+	uint32_t cmd;
+	int32_t arg;
+	int32_t arg2;
+	int32_t arg3;
+	int32_t arg4;
 } NetPacket;
 
 #define NETPKTBUFSIZE (4 + 4 + 4 + 4 + 4)
